♻️ refactor: 사용하지 않는 mock 파일 정리
