--- src/features/application/model/applicationMock.ts
@@ -1,269 +1,0 @@
-import type { ApplicationStats, ProjectApplication } from "./types"
-
-export const MOCK_STATS: ApplicationStats = {
-  totalMembers: 120,
-  completionRate: 64,
-  completedCount: 64,
-  pendingCount: 36,
-  rounds: [
-    { round: 1, applied: 78, total: 100 },
-    { round: 2, applied: 42, total: 58 },
-    { round: 3, applied: 24, total: 30 },
-  ],
-  topProjects: [
-    { name: "UMC", count: 22 },
-    { name: "유앰씨", count: 16 },
-    { name: "아무래도", count: 12 },
-    { name: "Figma", count: 9 },
-  ],
-  universities: [
-    { name: "서경대", applied: 0, total: 30 },
-    { name: "성신여대", applied: 0, total: 30 },
-    { name: "숭실대", applied: 0, total: 30 },
-    { name: "안양대", applied: 0, total: 30 },
-    { name: "한양대 ERICA", applied: 0, total: 30 },
-  ],
-  projectRounds: [
-    { name: "UMC", rounds: [80, 50, 20] },
-    { name: "유앰씨", rounds: [60, 40, 15] },
-    { name: "온라인모임", rounds: [55, 35, 10] },
-    { name: "승리기록", rounds: [50, 30, 10] },
-    { name: "슈어기", rounds: [45, 25, 8] },
-    { name: "순한기록", rounds: [40, 20, 5] },
-    { name: "플루에티", rounds: [35, 18, 5] },
-    { name: "토마토", rounds: [30, 15, 3] },
-    { name: "바나나요", rounds: [25, 12, 3] },
-    { name: "사운 글루", rounds: [20, 10, 2] },
-    { name: "챌린저스", rounds: [18, 8, 2] },
-    { name: "토마앰씨맛", rounds: [15, 7, 1] },
-    { name: "프로젝트 프로", rounds: [12, 5, 1] },
-    { name: "클루앤라이프 라라라", rounds: [10, 3, 1] },
-  ],
-}
-
-export const MOCK_PROJECTS: ProjectApplication[] = [
-  {
-    id: "1",
-    projectName: "프로젝트 명",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 2 },
-    feCount: { current: 3, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "2",
-    projectName: "Yoss",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 완료",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 4, total: 4 },
-    beCount: { current: 5, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "3",
-    projectName: "UMC_Web",
-    role: "plan",
-    challengerName: "벨라/황지원",
-    challengerUniversity: "중앙대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [
-      // 1차 (5명)
-      {
-        id: "3-1",
-        round: 1,
-        role: "design",
-        name: "김서연",
-        university: "한양대 ERICA",
-        status: "pass",
-        processedAt: { date: "04/22", time: "03:33" },
-        appliedAt: { date: "04/22", time: "01:30" },
-      },
-      {
-        id: "3-2",
-        round: 1,
-        role: "web",
-        name: "이준혁",
-        university: "숭실대",
-        status: "fail",
-        processedAt: { date: "04/22", time: "03:33" },
-        appliedAt: { date: "04/22", time: "01:38" },
-      },
-      {
-        id: "3-3",
-        round: 1,
-        role: "springboot",
-        name: "박하은",
-        university: "한양대 ERICA",
-        status: "pass",
-        processedAt: { date: "04/23", time: "10:50" },
-        appliedAt: { date: "04/20", time: "14:22" },
-      },
-      {
-        id: "3-4",
-        round: 1,
-        role: "springboot",
-        name: "최민수",
-        university: "성신여대",
-        status: "pending",
-        processedAt: null,
-        appliedAt: { date: "04/20", time: "16:45" },
-      },
-      {
-        id: "3-5",
-        round: 1,
-        role: "web",
-        name: "정다은",
-        university: "숭실대",
-        status: "fail",
-        processedAt: { date: "04/23", time: "11:05" },
-        appliedAt: { date: "04/21", time: "09:10" },
-      },
-      // 2차 (3명)
-      {
-        id: "3-6",
-        round: 2,
-        role: "springboot",
-        name: "한소희",
-        university: "한양대 ERICA",
-        status: "pass",
-        processedAt: { date: "05/01", time: "14:00" },
-        appliedAt: { date: "04/28", time: "10:00" },
-      },
-      {
-        id: "3-7",
-        round: 2,
-        role: "web",
-        name: "윤지호",
-        university: "안양대",
-        status: "pending",
-        processedAt: null,
-        appliedAt: { date: "04/28", time: "11:30" },
-      },
-      {
-        id: "3-8",
-        round: 2,
-        role: "springboot",
-        name: "강예린",
-        university: "숭실대",
-        status: "fail",
-        processedAt: { date: "05/01", time: "15:00" },
-        appliedAt: { date: "04/29", time: "09:00" },
-      },
-      // 3차 (2명)
-      {
-        id: "3-9",
-        round: 3,
-        role: "web",
-        name: "임도현",
-        university: "성신여대",
-        status: "pending",
-        processedAt: null,
-        appliedAt: { date: "05/10", time: "09:00" },
-      },
-      {
-        id: "3-10",
-        round: 3,
-        role: "springboot",
-        name: "조유나",
-        university: "한양대 ERICA",
-        status: "pending",
-        processedAt: null,
-        appliedAt: { date: "05/10", time: "11:00" },
-      },
-    ],
-  },
-  {
-    id: "4",
-    projectName: "모아모아",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "5",
-    projectName: "PACE",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 완료",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "6",
-    projectName: "타임투고",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "7",
-    projectName: "프로젝트",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "8",
-    projectName: "Nect",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "9",
-    projectName: "더트",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-  {
-    id: "10",
-    projectName: "타이프",
-    role: "plan",
-    challengerName: "닉네임/이름",
-    challengerUniversity: "한국대",
-    statusLabel: "모집 중",
-    designCount: { current: 1, total: 1 },
-    feCount: { current: 0, total: 4 },
-    beCount: { current: 0, total: 5 },
-    applicants: [],
-  },
-]
